grouped tests that cover the same method into groups
