+++ lib/airport.rb
@@ -1,5 +1,8 @@
 class Airport
 
+  require_relative 'plane'
+  require_relative 'weather'
+
   attr_reader :airport_array
   DEFAULT_CAPACITY = 5
 
@@ -11,7 +14,7 @@ class Airport
   def land(plane, forecast)
     forecast.weather_check == true ? (return full ? (raise "The airport is full") : @airport_array << plane) : "It is too stormy to take off"
   end
-  
+
   def take_off(plane, forecast)
     return forecast.weather_check ?
     (@airport_array.delete_at(airport_array.find_index(plane)); "The plane has successfully left the airport") : (raise "It is too stormy to take off")
@@ -21,24 +24,6 @@ class Airport
 
   def full
     @airport_array.length >= @capacity
-  end
-
-end
-
-class Plane
-
-end
-
-class Weather
-
-  attr_reader :weather
-
-  def initialize
-    @weather = (rand(1..10) == 10 ? :stormy : :sunny)
-  end
-
-  def weather_check
-    @weather == :sunny
   end
 
 end
