+++ lib/plane.rb
@@ -1,0 +1,3 @@
+class Plane
+
+end
